NotNull 제약조건 삭제

+++ domain-user/src/main/java/co/dalicious/domain/user/repository/UserRepository.java
@@ -1,12 +1,13 @@
 package co.dalicious.domain.user.repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import co.dalicious.domain.user.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, BigInteger> {
   Optional<User> findByEmail(String email);
 
   boolean existsUserByEmail(String email);
